+++ R/StatecCMP/R/zzz.R
@@ -1,4 +1,4 @@
 .onAttach <- function(libname, pkgname) {
-	packageStartupMessage("Statec functions and wrappers, version 0.0.6 hotfix, 2023-02-28.")
+	packageStartupMessage("Statec functions and wrappers, version 0.0.6.2, 2023-02-28.")
   packageStartupMessage("Maintained by Michel Geller (michel.geller@statec.etat.lu),\npreviously by Andrei V. Kostyrka (andrei.kostyrka@gmail.com).")
 }